Setup for ubuntu and minor tweaks

+++ create-links.sh
@@ -4,41 +4,42 @@
 echo ""
 echo "Setting up gitconfig links"
 gitconfig=~/.gitconfig
-rm "$gitconfig"
+[ -e "$gitconfig" ] && rm "$gitconfig"
 ln -s ~/dotfiles/git/gitconfig "$gitconfig"
 ls -al "$gitconfig"
 
 echo ""
 echo "Setting up gitignore links"
 gitignore=~/.gitignore
-rm "$gitignore"
+[ -e "$gitignore" ] && rm "$gitignore"
 ln -s ~/dotfiles/git/gitignore "$gitignore"
 ls -al "$gitignore"
 
 echo ""
 echo "Setting up zsh links"
 zsh=~/.zshrc
-rm "$zsh"
+[ -e "$zsh" ] && rm "$zsh"
 ln -s ~/dotfiles/zsh/zshrc "$zsh"
 ls -al "$zsh"
 
 echo ""
 echo "Setting up tmux links"
 tmux=~/.tmux.conf
-rm "$tmux"
+[ -e "$tmux" ] && rm "$tmux"
 ln -s ~/dotfiles/tmux/tmux.conf "$tmux"
 ls -al "$tmux"
 
 echo ""
 echo "Setting up alacritty links"
+mkdir -p ~/.config/alacritty/
 alacritty=~/.config/alacritty/alacritty.yml
-rm "$alacritty"
+[ -e "$alacritty" ] && rm "$alacritty"
 ln -s ~/dotfiles/alacritty/alacritty.yml "$alacritty"
 ls -al "$alacritty"
 
 echo ""
 echo "Setting up vim links"
 vim=~/.vimrc
-rm -f "$vim"
+[ -e "$vim" ] && rm "$vim"
 ln -s ~/dotfiles/vim/vimrc "$vim"
 ls -al "$vim"
